requirements.txt
pipwin==0.5.1
PyAudio @ file: PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pytest==7.0.0
requests==2.27.1
urllib3==1.26.8
PyAudio~=0.2.11
pydub~=0.25.1
simpleaudio~=1.0.4
PyYAML==6.0.1
